Separate out data widget
